refactor: shrink warning-sink infra to Consumer<String>; split logger-routing tests
Drop the WarningSink interface, default-sink field, and the explanatory
javadoc blocks in TypeConverter — collapse to a single ThreadLocal of
Consumer<String> with a 4-line warn() and a 5-line pushSink(). Same
behaviour, ~half the lines.

Move the duck-typed plugin-logger routing tests (and their plugin/logger
fixtures) out of YamlFileViaPluginTest into a dedicated
PluginLoggerRoutingTest, since they're a different concern than the
existing tests of the plugin-discovery mechanics.
