+++ src/test/java/org/avarion/yaml/YamlFileViaPluginTest.java
@@ -1,19 +1,11 @@
 package org.avarion.yaml;
 
-import org.avarion.yaml.testClasses.Material;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.io.TempDir;
 
 import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
-import java.util.ArrayList;
-import java.util.List;
-import java.util.logging.Handler;
-import java.util.logging.Level;
-import java.util.logging.LogRecord;
-import java.util.logging.Logger;
 
 import static org.assertj.core.api.Assertions.*;
 
@@ -116,197 +108,6 @@ class YamlFileViaPluginTest extends TestCommon {
         public File getDataFolder() {
             return null;
         }
-    }
-
-    // Plugin exposing both getDataFolder() and getLogger() — the shape Bukkit/Paper plugins have.
-    static class LoggingPlugin {
-        private final File dataFolder;
-        private final Logger logger;
-
-        public LoggingPlugin(File dataFolder, Logger logger) {
-            this.dataFolder = dataFolder;
-            this.logger = logger;
-        }
-
-        public File getDataFolder() {
-            return dataFolder;
-        }
-
-        public Logger getLogger() {
-            return logger;
-        }
-    }
-
-    @YamlFile(fileName = "config.yml", lenient = Leniency.LENIENT)
-    static class EnumListConfig extends YamlFileInterface {
-        @YamlKey("mats")
-        public List<Material> mats = List.of();
-    }
-
-    // Plugin with a getLogger() that's NOT public — must be ignored by the reflection lookup.
-    static class PrivateLoggerPlugin {
-        private final File dataFolder;
-        public PrivateLoggerPlugin(File dataFolder) { this.dataFolder = dataFolder; }
-        public File getDataFolder() { return dataFolder; }
-        @SuppressWarnings("unused")
-        private Logger getLogger() { return Logger.getLogger("private.should.be.ignored"); }
-    }
-
-    // Plugin whose getLogger() returns an object with no warn/warning method — must fall back to default.
-    static class WrongLoggerTypePlugin {
-        private final File dataFolder;
-        public WrongLoggerTypePlugin(File dataFolder) { this.dataFolder = dataFolder; }
-        public File getDataFolder() { return dataFolder; }
-        public String getLogger() { return "not a Logger"; }
-    }
-
-    // Plugin whose getLogger() returns null — must fall back to default.
-    static class NullLoggerPlugin {
-        private final File dataFolder;
-        public NullLoggerPlugin(File dataFolder) { this.dataFolder = dataFolder; }
-        public File getDataFolder() { return dataFolder; }
-        public Logger getLogger() { return null; }
-    }
-
-    // Logger whose warn(String) throws — verifies invokeQuietly swallows so the load doesn't blow up.
-    static class ThrowingLogger {
-        @SuppressWarnings("unused")
-        public void warn(String message) { throw new RuntimeException("boom"); }
-    }
-
-    static class ThrowingLoggerPlugin {
-        private final File dataFolder;
-        public ThrowingLoggerPlugin(File dataFolder) { this.dataFolder = dataFolder; }
-        public File getDataFolder() { return dataFolder; }
-        public ThrowingLogger getLogger() { return new ThrowingLogger(); }
-    }
-
-    // Fake SLF4J-style logger: exposes warn(String) but NOT warning(String). Records the calls.
-    static class Slf4jStyleLogger {
-        final List<String> warnings = new ArrayList<>();
-        @SuppressWarnings("unused")
-        public void warn(String message) { warnings.add(message); }
-    }
-
-    static class Slf4jPlugin {
-        private final File dataFolder;
-        private final Slf4jStyleLogger logger;
-        public Slf4jPlugin(File dataFolder, Slf4jStyleLogger logger) { this.dataFolder = dataFolder; this.logger = logger; }
-        public File getDataFolder() { return dataFolder; }
-        public Slf4jStyleLogger getLogger() { return logger; }
-    }
-
-    // Fake ALogger-style logger: exposes warning(String, Object...) varargs only.
-    static class VarargsStyleLogger {
-        final List<String> warnings = new ArrayList<>();
-        @SuppressWarnings("unused")
-        public void warning(String message, Object... args) { warnings.add(message); }
-    }
-
-    static class VarargsPlugin {
-        private final File dataFolder;
-        private final VarargsStyleLogger logger;
-        public VarargsPlugin(File dataFolder, VarargsStyleLogger logger) { this.dataFolder = dataFolder; this.logger = logger; }
-        public File getDataFolder() { return dataFolder; }
-        public VarargsStyleLogger getLogger() { return logger; }
-    }
-
-    @Test
-    void testLenientWarningGoesToPluginLoggerWhenLoadingViaPlugin() throws IOException {
-        File dataFolder = tempDir.toFile();
-        Files.writeString(new File(dataFolder, "config.yml").toPath(),
-                "mats:\n  - A\n  - NOT_A_MATERIAL\n  - B\n");
-
-        Logger pluginLogger = Logger.getLogger("test.plugin." + System.nanoTime());
-        pluginLogger.setUseParentHandlers(false);
-        List<LogRecord> pluginLogs = new ArrayList<>();
-        Handler handler = new Handler() {
-            @Override public void publish(LogRecord r) { pluginLogs.add(r); }
-            @Override public void flush() { /* no buffer */ }
-            @Override public void close() { /* no resources */ }
-        };
-        pluginLogger.addHandler(handler);
-
-        try {
-            LoggingPlugin plugin = new LoggingPlugin(dataFolder, pluginLogger);
-            EnumListConfig loaded = new EnumListConfig().load(plugin);
-            assertThat(loaded.mats).containsExactly(Material.A, Material.B);
-
-            // Warning routed to the plugin's logger, NOT the library default
-            assertThat(pluginLogs).hasSize(1);
-            assertThat(pluginLogs.get(0).getLevel()).isEqualTo(Level.WARNING);
-            assertThat(pluginLogs.get(0).getMessage()).contains("NOT_A_MATERIAL");
-            assertThat(logs).isEmpty();
-        } finally {
-            pluginLogger.removeHandler(handler);
-        }
-    }
-
-    @Test
-    void testNonPublicOrWrongTypeGetLoggerFallsBackToDefault() throws IOException {
-        File dataFolder = tempDir.toFile();
-        Files.writeString(new File(dataFolder, "config.yml").toPath(),
-                "mats:\n  - A\n  - NOT_A_MATERIAL\n");
-
-        // Private getLogger → reflection lookup (getMethod) must skip it; warning falls back to TypeConverter.LOG.
-        new EnumListConfig().load(new PrivateLoggerPlugin(dataFolder));
-        assertThat(logs).hasSize(1);
-        assertThat(logs.get(0).getMessage()).contains("NOT_A_MATERIAL");
-
-        logs.clear();
-
-        // Public getLogger() returning an object with no warn/warning method → fall back.
-        new EnumListConfig().load(new WrongLoggerTypePlugin(dataFolder));
-        assertThat(logs).hasSize(1);
-        assertThat(logs.get(0).getMessage()).contains("NOT_A_MATERIAL");
-
-        logs.clear();
-
-        // Public getLogger() returning null → fall back.
-        new EnumListConfig().load(new NullLoggerPlugin(dataFolder));
-        assertThat(logs).hasSize(1);
-        assertThat(logs.get(0).getMessage()).contains("NOT_A_MATERIAL");
-    }
-
-    @Test
-    void testThrowingPluginLoggerDoesNotBreakLoad() throws IOException {
-        File dataFolder = tempDir.toFile();
-        Files.writeString(new File(dataFolder, "config.yml").toPath(),
-                "mats:\n  - A\n  - NOT_A_MATERIAL\n  - B\n");
-
-        // Plugin's logger.warn(String) throws — invokeQuietly must swallow so the load completes.
-        EnumListConfig loaded = new EnumListConfig().load(new ThrowingLoggerPlugin(dataFolder));
-        assertThat(loaded.mats).containsExactly(Material.A, Material.B);
-    }
-
-    @Test
-    void testWarningRoutesToSlf4jStyleWarnMethod() throws IOException {
-        File dataFolder = tempDir.toFile();
-        Files.writeString(new File(dataFolder, "config.yml").toPath(),
-                "mats:\n  - A\n  - NOT_A_MATERIAL\n");
-
-        Slf4jStyleLogger fake = new Slf4jStyleLogger();
-        new EnumListConfig().load(new Slf4jPlugin(dataFolder, fake));
-
-        // Routed to the SLF4J-style warn(String); default sink stayed silent.
-        assertThat(fake.warnings).hasSize(1);
-        assertThat(fake.warnings.get(0)).contains("NOT_A_MATERIAL");
-        assertThat(logs).isEmpty();
-    }
-
-    @Test
-    void testWarningRoutesToVarargsStyleWarningMethod() throws IOException {
-        File dataFolder = tempDir.toFile();
-        Files.writeString(new File(dataFolder, "config.yml").toPath(),
-                "mats:\n  - A\n  - NOT_A_MATERIAL\n");
-
-        VarargsStyleLogger fake = new VarargsStyleLogger();
-        new EnumListConfig().load(new VarargsPlugin(dataFolder, fake));
-
-        // Routed to warning(String, Object...) with empty varargs; default sink stayed silent.
-        assertThat(fake.warnings).hasSize(1);
-        assertThat(fake.warnings.get(0)).contains("NOT_A_MATERIAL");
-        assertThat(logs).isEmpty();
     }
 
     @Test
